Organize osu! difficulty evaluators a bit better
